+++ changelog.md
@@ -8,12 +8,10 @@
   - Gold song names indicate a favorited song, and green song names indicate a currently playing song.
   - Each song entry (besides the currently playing one) has a play button so you can change songs anytime.
   - <c-FF0000>No, I will not be adding a search bar.</c>
+- Added `Minimal` as an option for `Button Mode`.
 - Retroactively rename `Playlist Mode` to `Constant Shuffle Mode` across as many front-facing areas as possible to avoid confusion with more recent features (see first bullet point).
 - Fix an issue where changing some settings didn't render a new song title card. Oops!
 - Fix an issue where blacklisting songs didn't update the name of the song chosen as the new menu loop. Oops!
-- Implemented soft incompatibility logic with a mod aiming to replace/redirect custom song directory locations.
-  - This should *not* affect songs in your custom songs folder.
-  - ...unless if you like to name those songs using only numbers. But that's your own problem to fix; this mod's been using raw file names for ages now.
 - Removed elasticity from MLR menu opening animations. <c-FF0000>Non-negotiable change.</c>
 - Moved repository URL to Codeberg.
 ## v1.8.4